% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\name{add_aggregator}
\alias{add_aggregator}
\title{add aggregator
note on regexp --> always case insensitive!
note overwrites an existing aggregator entry for the same dataset and column}
\usage{
add_aggregator(
  aggregator,
  dataset,
  column,
  source = column,
  default = NA,
  cast = "as.character",
  regexp = FALSE,
  func = NULL,
  args = NULL
)
}
\arguments{
\item{source}{single character column name or vector of column names if alternatives could be the source}

\item{args}{additional arguments to pass to func}
}
\description{
add aggregator
note on regexp --> always case insensitive!
note overwrites an existing aggregator entry for the same dataset and column
}
