% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_functions.R
\name{add_excel_sheet}
\alias{add_excel_sheet}
\title{internal function to add an excel sheet to a workbook}
\usage{
add_excel_sheet(
  wb,
  sheet_name,
  dataset,
  dbl_digits = 2,
  col_max_width = 75,
  int_format = "0",
  dbl_format = sprintf(sprintf("\%\%.\%df", dbl_digits), 0)
)
}
\arguments{
\item{wb}{the workbook object}

\item{sheet_name}{the name of the excel sheet}

\item{dataset}{the data frame(s)}

\item{dbl_digits}{how many digits to show for dbls (all are exported)}

\item{col_max_width}{maximum column width}

\item{int_format}{the excel formatting style for integers}

\item{dbl_format}{the excel formatting style for doubles (created automatically from the dbl_digits parameter)}
}
\description{
internal function to add an excel sheet to a workbook
}
