% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculate_gmean}
\alias{calculate_gmean}
\title{Geometric mean}
\usage{
calculate_gmean(x)
}
\arguments{
\item{x}{A numeric vector used to calculate the geometric mean}
}
\value{
The calculated geometric mean
}
\description{
The function calculate_gmean() is used to calculate geometric means
}
\keyword{internal}
