% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculate_gsd}
\alias{calculate_gsd}
\title{Standard deviation (geometric)}
\usage{
calculate_gsd(x)
}
\arguments{
\item{x}{A numeric values used to calculate the geometric standard deviation}
}
\value{
The calculated geometric standard deviation
}
\description{
The function calculate_gsd() is used to calculate geometric standard deviations
}
\keyword{internal}
