% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculate_gse}
\alias{calculate_gse}
\title{Standard error (geometric)}
\usage{
calculate_gse(x)
}
\arguments{
\item{x}{value to compute geometric standard error from}
}
\value{
The calculated geometric standard error
}
\description{
Define 'gse' to calculate geometic standard error
}
