% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratio_functions.R
\name{calculate_ratio_gsd}
\alias{calculate_ratio_gsd}
\title{Internal function to calculate standard deviation (geometric)}
\usage{
calculate_ratio_gsd(ratios)
}
\arguments{
\item{ratios}{A numeric values used to calculate the geometric standard deviation}
}
\value{
The calculated geometric standard deviation
}
\description{
The function \code{calculate_ratio_gsd()} is used to calculate geometric standard deviations
}
\keyword{internal}
