% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratio_functions.R
\name{calculate_ratio_sem}
\alias{calculate_ratio_sem}
\title{Internal function to calculate standard error}
\usage{
calculate_ratio_sem(ratios)
}
\arguments{
\item{ratios}{A numeric vector used to calculate a standard error}
}
\value{
The calculated standard error
}
\description{
The function \code{calculate_ratio_sem()} computes a regular standard error
}
\keyword{internal}
