% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{calculate_ratios}
\alias{calculate_ratios}
\title{ratio calculation fucntions}
\usage{
calculate_ratios(numerator, denominator, method)
}
\arguments{
\item{numerator}{xyz}

\item{denominator}{xyz}
}
\description{
ratio calculation fucntions
}
\details{
calculate ratios
}
\seealso{
Other ratio functions: 
\code{\link{calculate_ratios_se}()}
}
\concept{ratio functions}
\keyword{internal}
