% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratio_functions.R
\name{calculate_ratios_gmean}
\alias{calculate_ratios_gmean}
\title{Internal function to calculate geometric mean}
\usage{
calculate_ratios_gmean(ratios)
}
\arguments{
\item{ratios}{A numeric vector of ratios used to calculate the geometric mean}
}
\value{
The calculated geometric mean
}
\description{
The function \code{calculate_ratios_gmean()} is used to calculate geometric means
}
\keyword{internal}
