% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{calculate_ratios_se}
\alias{calculate_ratios_se}
\title{Calculate ratios se?}
\usage{
calculate_ratios_se(numerator, denominator)
}
\arguments{
\item{numerator}{xyz}

\item{denominator}{xyz}
}
\description{
Calculate ratios se?
}
\details{
\code{calculate_stuff}
}
\seealso{
Other ratio functions: 
\code{\link{calculate_ratios}()}
}
\concept{ratio functions}
\keyword{internal}
