% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratio_functions.R
\name{calculate_ratios_slope}
\alias{calculate_ratios_slope}
\title{Internal function for ratio_method \code{slope}}
\usage{
calculate_ratios_slope(x, y)
}
\arguments{
\item{x}{Vector of values used as ratio numerator}

\item{y}{Vector of values used as ratio denominator}
}
\value{
The calculated slope, an estimate of the ratio x/y
}
\description{
The function \code{calculate_ratios_slope()} is used to estimate the slope of x, y values used in a ratio
}
\details{
The slope is calculated from a linear regression model that is weighted by the numerator x, using \code{stats::lm(x ~ y + 0, weights = x)}
}
\keyword{internal}
