% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratio_functions.R
\name{calculate_ratios_weighted_sum}
\alias{calculate_ratios_weighted_sum}
\title{Internal function for ratio_method \code{weighted_sum}}
\usage{
calculate_ratios_weighted_sum(x, y)
}
\arguments{
\item{x}{A vector of values used as ratio numerator}

\item{y}{A vector of values used as ratio denominator}
}
\value{
The calculated ratio x/y
}
\description{
The function \code{calculate_ratios_weighted_sum()} is used to calculate ratios by weighted sums of x and y values
}
\details{
The weighing function ensures that each scan contributes equal weight to the ratio calculation,
i.e. scans with more ions in the Orbitrap do not contribute disproportionally to the total sum of x and y that is used to calculate x/y.
}
\keyword{internal}
