% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculate_se}
\alias{calculate_se}
\title{Standard error}
\usage{
calculate_se(x)
}
\arguments{
\item{x}{A numeric vector used to calculate a standard error}
}
\value{
The calculated standard error
}
\description{
Calculates a regular standard error
}
\keyword{internal}
