% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculate_weighted.vector.sum}
\alias{calculate_weighted.vector.sum}
\title{weighted.vector.sum}
\usage{
calculate_weighted.vector.sum(x, y)
}
\arguments{
\item{x}{???}

\item{y}{???}
}
\value{
???
}
\description{
???
}
