% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{factor_in_order}
\alias{factor_in_order}
\title{Create factor in order}
\usage{
factor_in_order(x)
}
\arguments{
\item{x}{vector or factor}
}
\value{
factor with levels in order of appearance
}
\description{
Helper function to create factors with levels in the order of data appearance. This is a simpler implementation of \code{\link[forcats:fct_inorder]{forcats::fct_inorder()}}
}
