% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logic_isox_file_interactions.R
\name{iso_read_isox_file}
\alias{iso_read_isox_file}
\title{Read .isox file}
\usage{
iso_read_isox_file(filepath)
}
\arguments{
\item{filepath}{Path to the .isox file}
}
\value{
A dataframe containing at minimum named 'filename', 'scan.no', 'time.min', 'compound', 'isotopolog', 'ions.incremental', 'tic', 'it.ms'
}
\description{
Read an IsoX output file into a tibble
}
\examples{
fpath <- system.file("extdata", "testfile_DualInlet_small.isox", package="isoorbi")
df <- iso_read_isox_file(filepath = fpath)

}
