% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logic_isox_common_core.R
\name{iso_simplify_isox_file}
\alias{iso_simplify_isox_file}
\title{Simplify IsoX output}
\usage{
iso_simplify_isox_file(dataset)
}
\arguments{
\item{dataset}{The loaded IsoX data that is to be simplified}
}
\value{
A simplified data frame with the columns: 'filename', 'scan.no', 'time.min', 'compound', 'isotopolog', 'ions.incremental', 'tic', 'it.ms'.
}
\description{
Keep only columns that are essential for isotopocule ratio analysis
}
\examples{
fpath <- system.file("extdata", "testfile_Flow_Exploration_small.isox", package="isoorbi")
df <- iso_read_isox_file(filepath = fpath) \%>\% iso_simplify_isox_file()

}
