% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoorbi-package.R
\docType{package}
\name{isoorbi-package}
\alias{isoorbi}
\alias{isoorbi-package}
\title{isoorbi: Process Orbitrap Isotopocule Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Read and process isotopocule data from an Orbitrap Isotope Solutions mass spectrometer. Citation: Kantnerova et al. (Nature Protocols, 2024).
}
\details{
Resources:
\itemize{
\item Website for the isoorbi package: \url{https://isoorbi.isoverse.org}
\item Package options: \link{orbi_options}
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://isoorbi.isoverse.org/}
  \item \url{https://github.com/isoverse/isoorbi}
  \item Report bugs at \url{https://github.com/isoverse/isoorbi/issues}
}

}
\author{
\strong{Maintainer}: Caj Neubauer \email{caj.neubauer@colorado.edu} (\href{https://orcid.org/0000-0002-5348-5609}{ORCID})

Authors:
\itemize{
  \item Caj Neubauer \email{caj.neubauer@colorado.edu} (\href{https://orcid.org/0000-0002-5348-5609}{ORCID})
  \item Sebastian Kopf \email{sebastian.kopf@colorado.edu} (\href{https://orcid.org/0000-0002-2044-0201}{ORCID})
  \item Kristýna Kantnerová \email{kristyna.kantnerova@colorado.edu} (\href{https://orcid.org/0000-0001-6259-3225}{ORCID})
}

}
