% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocks_functions.R
\name{orbi_adjust_block}
\alias{orbi_adjust_block}
\title{Manually adjust a block delimiter}
\usage{
orbi_adjust_block(
  dataset,
  block,
  filename = NULL,
  shift_start_time.min = NULL,
  shift_end_time.min = NULL,
  shift_start_scan.no = NULL,
  shift_end_scan.no = NULL,
  set_start_time.min = NULL,
  set_end_time.min = NULL,
  set_start_scan.no = NULL,
  set_end_scan.no = NULL
)
}
\arguments{
\item{dataset}{An aggregated dataset or a data frame of peaks (i.e. works directly after \code{\link[=orbi_identify_isotopocules]{orbi_identify_isotopocules()}} as well as with a tibble from \link[=orbi_get_data]{orbi_get_data(peaks = everything())} or when reading from an IsoX file)}

\item{block}{the block for which to adjust the start and/or end}

\item{filename}{needs to be specified only if the \code{dataset} has more than one \code{filename}}

\item{shift_start_time.min}{if provided, the start time of the block will be shifted by this many minutes (use negative numbers to shift back)}

\item{shift_end_time.min}{if provided, the end time of the block will be shifted by this many minutes (use negative numbers to shift back)}

\item{shift_start_scan.no}{if provided, the start of the block will be shifted by this many scans (use negative numbers to shift back)}

\item{shift_end_scan.no}{if provided, the end of the block will be shifted by this many scans (use negative numbers to shift back)}

\item{set_start_time.min}{if provided, sets the start time of the block as close as possible to this time}

\item{set_end_time.min}{if provided, sets the end time of the block as close as possible to this time}

\item{set_start_scan.no}{if provided, sets the start of the block to this scan number (scan must exist in the \code{dataset})}

\item{set_end_scan.no}{if provided, sets the end of the block to this scan number (scan must exist in the \code{dataset})}
}
\value{
see \code{\link[=orbi_adjust_blocks]{orbi_adjust_blocks()}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{orbi_adjust_block()} was renamed \code{\link[=orbi_adjust_blocks]{orbi_adjust_blocks()}} since it can now adjust several blocks at once
(and identifies the file with \code{in_filename} instead of \code{filename}).
}
