% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shotnoise_functions.R
\name{orbi_analyze_shot_noise}
\alias{orbi_analyze_shot_noise}
\title{Shot noise calculation}
\usage{
orbi_analyze_shot_noise(dataset, include_flagged_data = FALSE)
}
\arguments{
\item{dataset}{An aggregated dataset or a data frame of peaks (i.e. works directly after \code{\link[=orbi_identify_isotopocules]{orbi_identify_isotopocules()}} as well as with a tibble from \link[=orbi_get_data]{orbi_get_data(peaks = everything())} or when reading from an IsoX file)}

\item{include_flagged_data}{whether to include flagged data in the shot noise calculation (FALSE by default)}
}
\value{
The processed data frame with new columns: \code{n_effective_ions}, \code{ratio}, \code{ratio_rel_se.permil}, \code{shot_noise.permil}
}
\description{
This function computes the shot noise calculation.
}
\details{
Analyze shot noise

will calculate for all combinations of \code{filename}, \code{compound}, and \code{isotopocule} in the provided \code{dataset}
}
