% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dual_inlet_functions.R
\name{orbi_annotate_dualInlet}
\alias{orbi_annotate_dualInlet}
\title{orbi_annotate_dualInlet}
\usage{
orbi_annotate_dualInlet(data, annotations)
}
\arguments{
\item{data}{Filtered dual inlet data}

\item{annotations}{Annotation table provided by \code{orbi_define_dualInlet()}}
}
\value{
Returns an annotated data frame from dual inlet experiments
}
\description{
This function merges annotations between input data and dual inlet data
}
