% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{orbi_basepeak}
\alias{orbi_basepeak}
\title{Assign the base peak}
\usage{
orbi_basepeak(dataset, basepeak)
}
\arguments{
\item{dataset}{A tibble from a \code{IsoX} output. Needs to contain columns for \code{filename}, \code{compound}, \code{scan.no}, \code{isotopocule}, \code{ions.incremental}.}

\item{basepeak}{The isotopocule that gets assigned as base peak, i.e. the denominator to calculate ratios}
}
\value{
Input data frame plus two columns called \code{Basepeak} and \code{Basepeak.Ions}
}
\description{
\code{orbi_basepeak()} assigns one isotopocule in the data frame as the base peak
}
\examples{
fpath <- system.file("extdata", "testfile_Flow_Exploration_small.isox", package = "isoorbi")
df <- orbi_read_isox(filepath = fpath) \%>\%
                     orbi_simplify_isox() \%>\%
                     orbi_basepeak(basepeak = "M0")

}
