% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{orbi_calculate_gmean}
\alias{orbi_calculate_gmean}
\title{Geometric mean}
\usage{
orbi_calculate_gmean(x)
}
\arguments{
\item{x}{value to compute geometric mean from}
}
\value{
The calculated geometric mean
}
\description{
Define 'gmean' to calculate geometic mean
}
