% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{orbi_calculate_ratio}
\alias{orbi_calculate_ratio}
\title{Calculate ratio}
\usage{
orbi_calculate_ratio(
  peak1,
  peak2,
  ratio.method = c("mean", "sum", "median", "geometric.mean", "slope", "weighted.sum")
)
}
\arguments{
\item{peak1}{Peak 1 of interest}

\item{peak2}{Peak 2 of interest}

\item{ratio.method}{The desired method of computing ratio}
}
\value{
calculated ratio between peak1 and peak2 with selected ratio method
}
\description{
Ratio calculation between peaks of interest. NOTE: The formula used to average ratios matters! Do not simply use arithmetic mean.
}
