% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratio_functions.R
\name{orbi_calculate_ratios}
\alias{orbi_calculate_ratios}
\title{Calculate direct isotopocule ratios}
\usage{
orbi_calculate_ratios(dataset)
}
\arguments{
\item{dataset}{A data frame output after running \code{orbi_define_basepeak()}}
}
\value{
Returns a mutated dataset with \code{ratio} column added.
}
\description{
This function calculates isotopocule/base peak ratios for all isotopocules. It does not summarize or average the ratios in any way. For a summarizing version of this function, see \code{orbi_summarize_results()}.
}
