% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{orbi_calculate_results}
\alias{orbi_calculate_results}
\title{Generate the results table}
\usage{
orbi_calculate_results(dataset, ratio.method)
}
\arguments{
\item{dataset}{A processed tibble produced from \code{IsoX} output}

\item{ratio.method}{Method for computing the ratio; passed to \code{orbi_calculate_ratio()}}
}
\value{
Returns table containing \code{filename}, \code{compound},  \code{Basepeak}, \code{Isotopocule}, \code{Ratio}, \code{Ratio.SEM}, \code{relSE.permil}, \code{Shot.Noise.permil}, \code{No.of.Scans}, \code{Mins.to.1mio}
}
\description{
Contains the logic to generate the results table
}
\details{
Please note well: The formula used to calculate ion ratios matters! Do not simply use arithmetic mean.
The best option may depend on the type of data you are processing (e.g., MS1 versus M+1 fragmentation).

\strong{Description of options for \code{ratio.method}:}

\code{mean}: arithmetic mean of ratios from individual scans.

\code{sum}: sum of all ions of the numerator across all scans divided by the sum of all ions observed for the denominator across all scans.

\code{geometric.mean}: geometric mean of ratios from individual scans.

\code{slope}: The ratio is calculated using the slope obtained from a linear regression model that is weighted by the \verb{numerator x}, using \code{stats::lm(x ~ y + 0, weights = x)}.

\code{weighted.sum}: A derivative of the \code{sum} option. The weighing function ensures that each scan contributes equal weight to the ratio calculation,
i.e. scans with more ions in the Orbitrap do not contribute disproportionately to the total \code{sum} of \code{x} and \code{y} that is used to calculate \code{x/y}.
}
\examples{
fpath <- system.file("extdata", "testfile_Flow_Exploration_small.isox", package = "isoorbi")
df <- orbi_read_isox(filepath = fpath) \%>\%
      orbi_simplify_isox() \%>\% orbi_basepeak(basepeak = "M0")  \%>\%
      orbi_calculate_results(ratio.method = "sum")

}
