% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocks_functions.R
\name{orbi_define_block}
\alias{orbi_define_block}
\title{Manually define a block}
\usage{
orbi_define_block(
  dataset,
  start_time.min = NULL,
  end_time.min = NULL,
  start_scan.no = NULL,
  end_scan.no = NULL,
  sample_name = NULL
)
}
\arguments{
\item{dataset}{tibble with orbitrap data}

\item{start_time.min}{set the start time of the block}

\item{end_time.min}{set the end time of the block}

\item{start_scan.no}{set the start scan of the block}

\item{end_scan.no}{set the end scan of the block}

\item{sample_name}{if provided, will be used as the \code{sample_name} for the block}
}
\value{
A data frame (tibble) with block definition added. Any data that is not part of a block will be marked with the value of \code{orbi_get_settings("data_type_unused")}. Any previously applied segmentation will be discarded (\code{segment} column set to \code{NA}) to avoid unintended side effects.
}
\description{
Define a data block by either start and end time or start and end scan number.
Note that mannually defining blocks removes all block segmentation. Make sure to call \code{\link[=orbi_segment_blocks]{orbi_segment_blocks()}} \strong{after} finishing block definitions.
}
\details{
FIXME: complete description and parameters
}
