% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dual_inlet_functions.R
\name{orbi_dualInlet_annotate}
\alias{orbi_dualInlet_annotate}
\title{Merge dual inlet annotations with data}
\usage{
orbi_dualInlet_annotate(data, annotations)
}
\arguments{
\item{data}{Filtered dual inlet data}

\item{annotations}{Annotation table provided by \code{orbi_dualInlet_define()}}
}
\value{
Returns an annotated data frame from dual inlet experiments
}
\description{
This function merges annotations between input data and dual inlet data
}
