% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dual_inlet_functions.R
\name{orbi_dualInlet_define}
\alias{orbi_dualInlet_define}
\title{Define dual inlet experimental settings and data processing parameters}
\usage{
orbi_dualInlet_define(
  reference = "ref",
  sample = "smpl",
  number.of.blocks,
  switch.time,
  block.time,
  startup.time = 0,
  segments = 1
)
}
\arguments{
\item{reference}{Name of the solution in the reference syringe (default = "ref")}

\item{sample}{Name of the solution in the sample syringe (default = "smpl")}

\item{number.of.blocks}{Number of infusion blocks}

\item{switch.time}{Time needed for new solution to arrive after switching the valve}

\item{block.time}{Time duration of each infusion block in minutes}

\item{startup.time}{Optional: Time added in beginning for equilibration (in minutes, default = 0)}

\item{segments}{Optional: Each infusion block can be divided up into several segments (default = 1)}
}
\value{
Returns a data frame used to annotate data from dual inlet analysis
}
\description{
This function defines the experiment and ensure input has required information
}
