% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_functions.R
\name{orbi_export_data_to_excel}
\alias{orbi_export_data_to_excel}
\title{Export data frame to excel}
\usage{
orbi_export_data_to_excel(
  dataset,
  file,
  dbl_digits = 7,
  int_format = "0",
  dbl_format = sprintf(sprintf("\%\%.\%df", dbl_digits), 0)
)
}
\arguments{
\item{dataset}{data frame}

\item{file}{file path to export the file}

\item{dbl_digits}{how many digits to show for dbls (all are exported)}

\item{int_format}{the excel formatting style for integers}

\item{dbl_format}{the excel formatting style for doubles (created automatically from the dbl_digits parameter)}
}
\value{
returns dataset invisibly for use in pipes
}
\description{
This functions exports the final \code{dataset} into an Excel file.
}
