% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{orbi_filter_flagged_data}
\alias{orbi_filter_flagged_data}
\title{Filter out flagged data}
\usage{
orbi_filter_flagged_data(dataset)
}
\arguments{
\item{dataset}{a tibble with previously flagged data from \code{orbi_flat_satellite_peaks()}, \code{orbi_filter_weak_isotopocules()}, and/or \code{orbi_flag_outliers()}}
}
\value{
the
}
\description{
Filter out flagged data
}
