% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{orbi_filter_isox}
\alias{orbi_filter_isox}
\title{Basic  filtering of IsoX data}
\usage{
orbi_filter_isox(
  dataset,
  isotopocules,
  base_peak,
  filenames,
  compounds,
  time_min,
  time_max
)
}
\arguments{
\item{dataset}{The data frame to be filtered}

\item{isotopocules}{Vector of isotopocules to select}

\item{base_peak}{Name of base peak to select}

\item{filenames}{Vector of filenames to select}

\item{compounds}{Vector of compounds files to select}

\item{time_min}{Minimum time in minutes}

\item{time_max}{Maximum time in minutes}
}
\value{
Filtered tibble
}
\description{
A basic generic filter for filenames, isotopocules, compounds and time ranges
}
