% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{orbi_filter_satellitePeaks}
\alias{orbi_filter_satellitePeaks}
\title{remove.satellitePeaks}
\usage{
orbi_filter_satellitePeaks(df)
}
\arguments{
\item{df}{Simplified IsoX file to have satellites removed from}
}
\value{
IsoX file with satellites removed
}
\description{
Remove minor signals (e.g., satellite peaks) that were reported by IsoX
}
