% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{orbi_filter_scan_intensity}
\alias{orbi_filter_scan_intensity}
\title{Function replaced by \code{orbi_flag_outliers()}}
\usage{
orbi_filter_scan_intensity(..., outlier_percent)
}
\description{
Function replaced by \code{orbi_flag_outliers()}
}
