% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{orbi_filter_TICxIT}
\alias{orbi_filter_TICxIT}
\title{Filter to remove extreme scans}
\usage{
orbi_filter_TICxIT(dataset, truncate_extreme)
}
\arguments{
\item{dataset}{Simplified IsoX file to have \code{TICxIT} outliers removed}

\item{truncate_extreme}{A number between 0 and 10. Remove this percentage of scans based on TIC multiplied by injection time.}
}
\value{
Filtered data frame (tibble)
}
\description{
Remove extremely high and low intense scans based on TIC x Injection time with \code{orbi_filter_TICxIT()}
}
\details{
The \code{orbi_filter_TICxIT()} function removes scans that are outliers. TIC multiplied by injection time serves as an estimate for the number of ions in the Orbitrap.
The filter is a basic truncation that removed x \% of scans with maximal \strong{and} x \% of scans with minimal ion estimates. Grouping is by \code{filename} and \code{compound}.
}
\examples{
fpath <- system.file("extdata", "testfile_Flow_Exploration_small.isox", package = "isoorbi")
df <- orbi_read_isox(filepath = fpath) \%>\%
orbi_simplify_isox() \%>\%
orbi_filter_TICxIT(truncate_extreme = 1)

}
