% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading_functions.R
\name{orbi_find_isox}
\alias{orbi_find_isox}
\title{Find isox files}
\usage{
orbi_find_isox(folder, recursive = TRUE)
}
\arguments{
\item{folder}{path to a folder with isox files}

\item{recursive}{whether to find files recursively}
}
\description{
Finds all .isox files in a folder.
}
\examples{

# all .isox files provided with the isoorbi package
orbi_find_isox(system.file("extdata", package = "isoorbi"))

}
