% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_raw.R
\name{orbi_find_raw}
\alias{orbi_find_raw}
\title{Find raw files}
\usage{
orbi_find_raw(folder, recursive = TRUE)
}
\arguments{
\item{folder}{path to a folder with raw files}

\item{recursive}{whether to find files recursively}
}
\description{
Finds all .raw files in a folder.
}
\examples{

# all .raw files provided with the isoorbi package
orbi_find_raw(system.file("extdata", package = "isoorbi"))

}
