% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{orbi_flag_outliers}
\alias{orbi_flag_outliers}
\title{Flag outlier scans}
\usage{
orbi_flag_outliers(dataset, intensity_window)
}
\arguments{
\item{dataset}{Simplified IsoX dataset to have outliers flagged}

\item{intensity_window}{flags extremely high and low intense scans based on TIC x injection time (i.e., ion intensity). Provide a vector with 2 numbers \code{c(x,y)} filtering the lowest x percent intensities and highest y percent intensities.}
}
\value{
Filtered tibble
}
\description{
The function \code{orbi_flag_outliers()} flags outliers. Grouping is by columns \code{filename} and \code{compound}.
}
\details{
Function is intended to flag scans that are outliers. TIC multiplied by injection time serves as an estimate for the number of ions in the Orbitrap.

The input \code{dataset} is expected to have at least these 8 columns: \code{filename}, \code{scan.no}, \code{time.min}, \code{compound}, \code{isotopocule}, \code{ions.incremental}, \code{tic}, \code{it.ms}.
}
\examples{
fpath <- system.file("extdata", "testfile_flow.isox", package = "isoorbi")
df <-
  orbi_read_isox(file = fpath) |>
  orbi_simplify_isox() |>
  orbi_flag_outliers(intensity_window = c(1,99))

}
