% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{orbi_get_isotopocule_coverage}
\alias{orbi_get_isotopocule_coverage}
\title{Calculate isotopocule coverage}
\usage{
orbi_get_isotopocule_coverage(dataset)
}
\arguments{
\item{dataset}{A data frame or tibble produced from IsoX data}
}
\value{
summary data frame
}
\description{
Calculate which stretches of the data have data for which isotopocules. This function is usually used indicrectly by \code{orbi_plot_isotopocule_coverage()} but can be called directly to investigate isotopocule coverage.
}
