% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/problems.R
\name{orbi_get_problems}
\alias{orbi_get_problems}
\title{Retrieve parsing problems}
\usage{
orbi_get_problems(obj)
}
\arguments{
\item{obj}{data object that holds problems information}
}
\value{
tibble data frame with a list of problems encountered during
}
\description{
This function retrieves parsing problems encountered during the reading and processing of files.
}
\examples{

orbi_find_raw(system.file("extdata", package = "isoorbi"))[1] |>
  orbi_read_raw() |>
  orbi_get_problems()

}
