% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options-package.R
\name{orbi_get_settings}
\alias{orbi_get_settings}
\title{Get all isoorbi package settings}
\usage{
orbi_get_settings(pattern = NULL)
}
\arguments{
\item{pattern}{passed on to \code{\link[=orbi_get_options]{orbi_get_options()}}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{orbi_get_settings()} was renamed \code{orbi_get_options()} as part of \code{isoorbi} switching from 'settings' to 'options' to be consistent with base R naming conventions
}
