% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R, R/processing.R
\name{orbi_plot_isotopocule_coverage}
\alias{orbi_plot_isotopocule_coverage}
\alias{orbi_get_isotopocule_coverage}
\title{Isotopocule coverage}
\usage{
orbi_plot_isotopocule_coverage(
  dataset,
  isotopocules = c(),
  x = c("scan.no", "time.min"),
  x_breaks = NULL,
  n_x_breaks = 5,
  short_time_labels = FALSE,
  add_data_blocks = TRUE,
  colors = c("#7570B3", "#E6AB02", "#66A61E", "#A6761D", "#D95F02", "#1B9E77", "#E7298A",
    "#666666", "#BBBBBB")
)

orbi_get_isotopocule_coverage(dataset)
}
\arguments{
\item{dataset}{a data frame or aggregated dataset with satellite peaks already identified (i.e. after \code{\link[=orbi_flag_satellite_peaks]{orbi_flag_satellite_peaks()}})}

\item{isotopocules}{which isotopocules to visualize, if none provided will visualize all (this may take a long time or even crash your R session if there are too many isotopocules in the data set)}

\item{x}{x-axis column for the plot, either "time.min" or "scan.no", default is "scan.no"}

\item{x_breaks}{what breaks to use for the x axis. By default (\code{NULL}) these are pretty breaks for scan numbers or pretty time intervals for time (which is labeled as a duration, e.g. \verb{1:30 min}), provide breaks to make more specific tickmarks. Use either \code{x_breaks} or \code{n_x_breaks}, not both.}

\item{n_x_breaks}{the desired number of x axis breaks when using the default pretty breaks (\code{x_breaks = NULL}), default: \code{5}. Use either \code{x_breaks} or \code{n_x_breaks}, not both.}

\item{short_time_labels}{whether to use compact time axis labels with no space between value and unit and abbreviated units (\code{hr}, \code{m}, \code{s}), e.g. \verb{1:30m} instead of \verb{1:30 min}. Only relevant for a time based x axis (\code{x = "time.min"}).}

\item{add_data_blocks}{add highlight for data blocks if there are any block definitions in the dataset (uses \code{\link[=orbi_add_blocks_to_plot]{orbi_add_blocks_to_plot()}}). To add blocks manually, set \code{add_data_blocks = FALSE} and manually call the \code{orbi_add_blocks_to_plot()} function afterwards.}

\item{colors}{the fill colors for the isotopocules that carry peak flags, one per flag combination encountered in the data (recycled if there are more combinations than colors). Isotopocules without any flags are always shown in black.}
}
\value{
a ggplot object

summary data frame
}
\description{
The coverage of each isotopcule across scans/time is an important indicator for data completeness. These functions provide ways to summarize and visualize the isotopocule coverage in a dataset.
}
\section{Functions}{
\itemize{
\item \code{orbi_plot_isotopocule_coverage()}: visualizes isotope coverage. Detected isotopocules are shown by their peak flags - those without any flags in black, those carrying flags in the \code{colors} provided. Weak isotopocules (if previously defined by \code{\link[=orbi_flag_weak_isotopocules]{orbi_flag_weak_isotopocules()}}) are highlighted in red.

\item \code{orbi_get_isotopocule_coverage()}: calculates which stretches of the data have data for which isotopocules. This function is usually used indicrectly by \code{\link[=orbi_plot_isotopocule_coverage]{orbi_plot_isotopocule_coverage()}} but can be called directly to investigate isotopocule coverage.

}}
