% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{orbi_plot_isotopocule_coverage}
\alias{orbi_plot_isotopocule_coverage}
\title{Plot isotopocule coverage}
\usage{
orbi_plot_isotopocule_coverage(
  dataset,
  isotopocules = c(),
  x = c("scan.no", "time.min"),
  add_blocks = TRUE
)
}
\arguments{
\item{dataset}{isox data}

\item{isotopocules}{which isotopocules to visualize, if none provided will visualize all (this may take a long time or even crash your R session if there are too many isotopocules in the data set)}

\item{x}{x-axis column for the plot, either "time.min" or "scan.no"}

\item{add_blocks}{add blocks if there are any block definitions in the dataset}
}
\description{
Plot isotopocule coverage
}
