% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocks_functions.R
\name{orbi_segment_blocks}
\alias{orbi_segment_blocks}
\title{Segment blocks}
\usage{
orbi_segment_blocks(dataset)
}
\arguments{
\item{dataset}{tibble produced by \code{\link[=orbi_define_blocks_for_dual_inlet]{orbi_define_blocks_for_dual_inlet()}}}
}
\description{
Segment blocks
}
