% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{orbi_set_settings}
\alias{orbi_set_settings}
\title{Set package settings}
\usage{
orbi_set_settings(
  di_ref_name = "ref",
  di_sample_name = "sam",
  data_type_data = "data",
  data_type_startup = "startup",
  data_type_changeover = "changeover",
  data_type_unused = "unused",
  reset_all = FALSE
)
}
\arguments{
\item{di_ref_name}{the text label for dual inlet reference blocks}

\item{di_sample_name}{the text label for dual inlet sample blocks}

\item{data_type_data}{the text used to flag raw data as actually being data}

\item{data_type_startup}{the text used to flag raw data as being part of the startup}

\item{data_type_changeover}{the text used to flag raw data as being part of a changeover}

\item{data_type_unused}{the text used to flag raw data as being unused}

\item{reset_all}{if set to TRUE, will reset all settings back to their defaults}
}
\value{
invisible list of all settings (see \code{\link[=orbi_get_settings]{orbi_get_settings()}})
}
\description{
Use this function to change the default package settings. When calling this function, only specify the settings you want to change, everything else will remain unchanged. The default value for each parameter is what the package uses by default for each setting.
}
\details{
FIXME: needs documentation completion
FIXME: needs tests to change settings and then get the value back
}
