% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\name{orbi_start_aggregator}
\alias{orbi_start_aggregator}
\title{start the aggregator for a dataset
defines the unique ID field}
\usage{
orbi_start_aggregator(
  dataset,
  uid_source,
  cast,
  regexp = FALSE,
  func = NULL,
  args = NULL
)
}
\description{
start the aggregator for a dataset
defines the unique ID field
}
