% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove.rare}
\alias{remove.rare}
\title{Remove rare isotopocules}
\usage{
remove.rare(data, min_percent)
}
\arguments{
\item{data}{Simplified IsoX data to be processed}

\item{min_percent}{Set threshold. Isotopocule must be observed in at least  x percent of scans}
}
\value{
Returns data frame with inconsistent isotopocules removed
}
\description{
Remove isotopocules that are not consistently detected across scans
}
