% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculate_gsd}
\alias{calculate_gsd}
\title{Standard deviation (geometric)}
\usage{
calculate_gsd(x)
}
\arguments{
\item{x}{value to compute geometric standard deviation from}
}
\value{
The calculated geometric standard deviation
}
\description{
Define 'gsd' to calculate geometic standard deviation
}
