% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculate_gse}
\alias{calculate_gse}
\title{Standard error (geometric)}
\usage{
calculate_gse(x)
}
\arguments{
\item{x}{A vector of values used to calculate geometric standard errors}
}
\value{
The calculated geometric standard error
}
\description{
The function calculate_gse() is used to calculate geometric standard errors
}
\keyword{internal}
