% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculate_se}
\alias{calculate_se}
\title{Standard error}
\usage{
calculate_se(x)
}
\arguments{
\item{x}{Vector to compute standard error from}
}
\value{
Standard error
}
\description{
Calculates standard error
}
