% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculate_slope}
\alias{calculate_slope}
\title{slope}
\usage{
calculate_slope(x, y)
}
\arguments{
\item{x}{???}

\item{y}{???}
}
\value{
The slope
}
\description{
Define 'slope' to calculate geometic mean
}
