% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoorbi.R
\docType{package}
\name{isoorbi-package}
\alias{isoorbi}
\alias{isoorbi-package}
\title{isoorbi: Process Orbitrap Iso Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

The goal of 'isoorbi' is to read and process isotopocule measurements from an Orbitrap Iso mass spectrometers. It expects '.isox' data produced by IsoX as input.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/isoverse/isoorbi}
  \item Report bugs at \url{https://github.com/isoverse/isoorbi/issues}
}

}
\author{
\strong{Maintainer}: Caj Neubauer \email{caj.neubauer@colorado.edu} (\href{https://orcid.org/0000-0002-5348-5609}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Sebastian Kopf \email{sebastian.kopf@colorado.edu} (\href{https://orcid.org/0000-0002-2044-0201}{ORCID})
  \item Kristýna Kantnerová \email{kristyna.kantnerova@colorado.edu} (\href{https://orcid.org/0000-0001-6259-3225}{ORCID})
}

}
\keyword{internal}
