% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\name{orbi_aggregate_raw}
\alias{orbi_aggregate_raw}
\title{aggregate data from raw files}
\usage{
orbi_aggregate_raw(
  files_data,
  aggregator = orbi_get_option("raw_aggregator"),
  show_progress = rlang::is_interactive(),
  show_problems = TRUE
)
}
\arguments{
\item{files_data}{the files read in by orbi_read_raw}
}
\description{
aggregate data from raw files
}
