% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{orbi_calculate_output}
\alias{orbi_calculate_output}
\title{Calculate output table}
\usage{
orbi_calculate_output(dataset, ratio.method)
}
\arguments{
\item{dataset}{A processed data frame from IsoX output}

\item{ratio.method}{Method  for computing the ratio; passed to 'iso_calculate_ratio()'}
}
\value{
Returns table containing \code{filename}, \code{compound},  \code{Basepeak}, \code{Isotopocule}, \code{Ratio}, \code{Ratio.SEM}, \code{relSE.permil}, \code{Shot.Noise.permil}, \code{No.of.Scans}, \code{Mins.to.1mio}
}
\description{
Contains the logic to generate the output table
}
\details{
Description of options for ratio.method:

"mean": arithmetic mean of ratios from individual scans.

"sum": sum of all ions of the numerator across all scans divided by the sum of all ions observed for the denominator across all scans.

"geometric.mean": geometric mean of ratios from individual scans.

"slope": The ratio is calculated using the slope from a linear model that is weighted by the numerator x, using stats::lm(x ~ y + 0, weights = x).

"weighted.sum": The weighing function ensures that each scan contributes equal weight to the ratio calculation,
i.e. scans with more ions in the Orbitrap do not contribute disproportionally to the total sum of x and y that is used to calculate x/y.
}
\examples{
fpath <- system.file("extdata", "testfile_Flow_Exploration_small.isox", package="isoorbi")
df <- orbi_read_isox(filepath = fpath) \%>\%
      orbi_simplify_isox() \%>\% orbi_basepeak(basepeak = "M0")  \%>\%
      orbi_calculate_output(ratio.method = "sum")

}
