% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratio_functions.R
\name{orbi_calculate_ratios}
\alias{orbi_calculate_ratios}
\title{Calculate isotopocule ratios}
\usage{
orbi_calculate_ratios(...)
}
\arguments{
\item{...}{parameters passed on to new function orbi_calculate_ratio()}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} This function was renamed to \code{\link[=orbi_calculate_ratio]{orbi_calculate_ratio()}} to better reflect what it does.
}
