% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{orbi_calculate_results}
\alias{orbi_calculate_results}
\title{Generate the results table}
\usage{
orbi_calculate_results(dataset, ratio_method)
}
\arguments{
\item{dataset}{A processed tibble produced from \code{IsoX} output}

\item{ratio_method}{Method for computing the ratio; passed to \code{orbi_calculate_ratio()}}
}
\value{
Returns a results table containing \code{filename}, \code{compound},  \code{basepeak}, \code{Isotopocule}, \code{Ratio}, \code{ratio_sem}, \code{ratio_relative_sem_permil}, \code{shot_noise_permil}, \code{No.of.Scans}, \code{minutes_to_1e6_ions}
}
\description{
Contains the logic to generate the results table
}
\details{
\strong{Description of the output columns:}
\itemize{
\item \code{basepeak}: Isotopocule used as denominator in ratio calculation.
\item \code{isotopocule}: Isotopocule used as numerator in ratio calculation.
\item \code{ratio_sem}: Standard error of the mean for the ratio
\item \code{number_of_scans}: Number of scans used for the final ratio calculation
\item \code{minutes_to_1e6_ions}: Time in minutes it would take to observe 1 million ions of the \code{isotopocule} used as numerator of the ratio calculation.
\item \code{shot_noise_permil}: Estimate of the shot noise (more correctly thermal noise) of the reported ratio in permil.
\item \code{ratio_relative_sem_permil}: Relative standard error of the reported ratio in permil
}

\strong{Description of options for \code{ratio_method}:}

Please note well: The formula used to calculate ion ratios matters! Do not simply use arithmetic mean.
The best option may depend on the type of data you are processing (e.g., MS1 versus M+1 fragmentation).
\itemize{
\item \code{mean}: arithmetic mean of ratios from individual scans.
\item \code{sum}: sum of all ions of the numerator across all scans divided by the sum of all ions observed for the denominator across all scans.
\item \code{geometric.mean}: geometric mean of ratios from individual scans.
\item \code{slope}: The ratio is calculated using the slope obtained from a linear regression model that is weighted by the \verb{numerator x}, using \code{stats::lm(x ~ y + 0, weights = x)}.
\item \code{weighted.sum}: A derivative of the \code{sum} option. The weighing function ensures that each scan contributes equal weight to the ratio calculation,
i.e. scans with more ions in the Orbitrap do not contribute disproportionately to the total \code{sum} of \code{x} and \code{y} that is used to calculate \code{x/y}.
}
}
\examples{
fpath <- system.file("extdata", "testfile_Flow_Exploration_small.isox", package = "isoorbi")
df <- orbi_read_isox(filepath = fpath) \%>\%
      orbi_simplify_isox() \%>\% orbi_define_basepeak(base_peak = "M0")  \%>\%
      orbi_calculate_results(ratio_method = "sum")

}
