% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{orbi_default_theme}
\alias{orbi_default_theme}
\title{Default isoorbi plotting theme}
\usage{
orbi_default_theme(text_size = 16, facet_text_size = 20)
}
\arguments{
\item{text_size}{a font size for text}

\item{facet_text_size}{a font size for facet text}
}
\value{
ggplot theme object
}
\description{
Default isoorbi plotting theme
}
