% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{orbi_filter_flagged_data}
\alias{orbi_filter_flagged_data}
\title{Filter out flagged data}
\usage{
orbi_filter_flagged_data(dataset)
}
\arguments{
\item{dataset}{a tibble with previously flagged data from \code{orbi_flag_satellite_peaks()}, \code{orbi_flag_weak_isotopocules()}, and/or \code{orbi_flag_outliers()}}
}
\value{
a dataset with the flagged data filtered out
}
\description{
This function filters out data that have been previously flagged using functions \code{orbi_flag_satellite_peaks()}, \code{orbi_flag_weak_isotopocules()}, and/or \code{orbi_flag_outliers()}. Note that this function is no longer necessary to call explicitly as \code{orbi_analyze_shot_noise()} and \code{orbi_summarize_results()} automatically exclude flagged data.
}
