% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{orbi_filter_isox}
\alias{orbi_filter_isox}
\title{orbi_filter_isox}
\usage{
orbi_filter_isox(
  dataset,
  isotopocules,
  base_peak,
  isox_files,
  compounds,
  time_min,
  time_max
)
}
\arguments{
\item{dataset}{The file.tsv to be filtered}

\item{isotopocules}{Vector of isotopocules to select}

\item{base_peak}{Name of base peak to select}

\item{isox_files}{Vector of isox files to select}

\item{compounds}{Vector of compounds files to select}

\item{time_min}{Minimum time in minutes}

\item{time_max}{Maximum time in minutes}
}
\value{
Filtered tibble
}
\description{
A basic filter (should be generalized for most use cases)
}
