% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_isox.R
\name{orbi_filter_isox}
\alias{orbi_filter_isox}
\title{Filter isox files}
\usage{
orbi_filter_isox(...)
}
\arguments{
\item{...}{arguments passed on to \code{\link[=orbi_filter_files]{orbi_filter_files()}}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
\code{\link[=orbi_filter_isox]{orbi_filter_isox()}} was renamed \code{\link[=orbi_filter_files]{orbi_filter_files()}} to incorporate its wider scope for filtering both isox and raw files data.
}
