% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{orbi_filter_isox}
\alias{orbi_filter_isox}
\title{Basic filtering of IsoX data}
\usage{
orbi_filter_isox(
  dataset,
  filenames,
  isotopocules,
  compounds,
  time_min,
  time_max
)
}
\arguments{
\item{dataset}{The data frame to be filtered}

\item{filenames}{Vector of file names to select}

\item{isotopocules}{Vector of isotopocules to select}

\item{compounds}{Vector of compounds files to select}

\item{time_min}{Minimum time in minutes}

\item{time_max}{Maximum time in minutes}
}
\value{
Filtered data frame (tibble)
}
\description{
A basic filter function \code{orbi_filter_isox()} for file names, isotopocules, compounds and time ranges
}
\examples{
fpath <- system.file("extdata", "testfile_Flow_Exploration_small.isox", package = "isoorbi")
df <- orbi_read_isox(filepath = fpath) \%>\%
orbi_simplify_isox() \%>\%
orbi_filter_isox(filenames = c("s3744"),
compounds = c("HSO4-"),
isotopocules = c("M0"),
time_min = 0,
time_max = 1)

}
