% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{orbi_filter_isox}
\alias{orbi_filter_isox}
\title{Basic generic filter for IsoX data}
\usage{
orbi_filter_isox(
  dataset,
  filenames,
  compounds,
  isotopocules,
  time_min,
  time_max
)
}
\arguments{
\item{dataset}{The IsoX data to be filtered}

\item{filenames}{Vector of file names to keep; also accepts \code{"all"}}

\item{compounds}{Vector of compounds to keep; also accepts \code{"all"}}

\item{isotopocules}{Vector of isotopocules to keep; also accepts \code{"all"}}

\item{time_min}{Minimum retention time in minutes (\code{time.min})}

\item{time_max}{Maximum retention time in minutes (\code{time.min})}
}
\value{
Filtered tibble
}
\description{
A basic filter function \code{orbi_filter_isox()} for file names, isotopocules, compounds and time ranges
}
\examples{
fpath <- system.file("extdata", "testfile_Flow_Exploration_small.isox", package = "isoorbi")
df <- orbi_read_isox(filepath = fpath) \%>\%
orbi_simplify_isox() \%>\%
orbi_filter_isox(filenames = c("s3744"),
compounds = c("HSO4-"),
isotopocules = "M0",
time_min = "all",
time_max = "all")

}
