% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{orbi_filter_satellite_peaks}
\alias{orbi_filter_satellite_peaks}
\title{Function replaced by \code{orbi_flag_satellite_peaks()}}
\usage{
orbi_filter_satellite_peaks(...)
}
\arguments{
\item{...}{parameters passed on to the new function \code{\link[=orbi_flag_satellite_peaks]{orbi_flag_satellite_peaks()}}.}
}
\description{
Function replaced by \code{orbi_flag_satellite_peaks()}
}
