% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{orbi_filter_satellitePeaks}
\alias{orbi_filter_satellitePeaks}
\title{Filter satellite peaks}
\usage{
orbi_filter_satellitePeaks(df)
}
\arguments{
\item{df}{Simplified IsoX file produced by orbi_simplify_isox()}
}
\value{
Filtered tibble
}
\description{
Remove minor signals (e.g., satellite peaks) that were reported by IsoX
}
