% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{orbi_filter_satellitePeaks}
\alias{orbi_filter_satellitePeaks}
\title{Filter satellite peaks}
\usage{
orbi_filter_satellitePeaks(df)
}
\arguments{
\item{df}{A data frame or tibble produced from IsoX data by orbi_simplify_isox()}
}
\value{
Filtered tibble
}
\description{
Remove minor signals (e.g., satellite peaks) that were reported by IsoX
}
\details{
The orbi_filter_satellitePeaks() function removes minor signals for an isotopocule that have been reported by IsoX.
These are often small 'satellite peaks' generated by the Fourier transform.
If there are unusual signals this may indicate that the m/z and tolerance setting used for processing .raw files with IsoX were incorrect.
}
