% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{orbi_filter_scan_intensity}
\alias{orbi_filter_scan_intensity}
\title{Function replaced by \code{orbi_flag_outliers()}}
\usage{
orbi_filter_scan_intensity(..., outlier_percent)
}
\arguments{
\item{...}{parameters passed on to the new function \code{\link[=orbi_flag_outliers]{orbi_flag_outliers()}}.}

\item{outlier_percent}{outlier_percent needs to be between 0 and 10, flags extreme scans based on TIC x injection time (i.e., ion intensity)}
}
\description{
Function replaced by \code{orbi_flag_outliers()}
}
