% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{orbi_filter_TICxIT}
\alias{orbi_filter_TICxIT}
\title{Filter by TIC x Injection Time}
\usage{
orbi_filter_TICxIT(df, truncate_extreme)
}
\arguments{
\item{df}{Simplified IsoX file to have 'ticxit' removed}

\item{truncate_extreme}{Remove extreme scans (peercentage) based on TIC x Injection time}
}
\value{
Filtered tibble
}
\description{
Remove extreme scans based on TIC x Injection time
}
