% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{orbi_filter_TICxIT}
\alias{orbi_filter_TICxIT}
\title{Filter by TIC x Injection Time}
\usage{
orbi_filter_TICxIT(df, truncate_extreme)
}
\arguments{
\item{df}{Simplified IsoX file to have 'ticxit' outliers removed}

\item{truncate_extreme}{Remove extreme scans (percentage) based on TIC x Injection time}
}
\value{
Filtered data frame
}
\description{
Remove extreme scans based on TIC x Injection time
}
\details{
The orbi_filter_TICxIT() function removes scans that are outliers. TIC x it.ms serves as a proxy for the number of ions in the Orbitrap. The filter is a basic truncation that removed x \% of scans with maximal or minimal ion estimates.
}
