% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{orbi_filter_TICxIT}
\alias{orbi_filter_TICxIT}
\title{Filter by TIC x Injection Time}
\usage{
orbi_filter_TICxIT(df, extreme_min)
}
\arguments{
\item{df}{Simplified IsoX file to have 'ticxit' removed}

\item{extreme_min}{???}
}
\value{
Returns filtered IsoX file based of input value
}
\description{
Remove extreme scans based on TIC x Injection time
}
