% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{orbi_filter_weak}
\alias{orbi_filter_weak}
\title{Filter weak isotopocules}
\usage{
orbi_filter_weak(data, min_percent)
}
\arguments{
\item{data}{Simplified IsoX data to be processed}

\item{min_percent}{Set threshold. Isotopocule must be observed in at least  x percent of scans}
}
\value{
Filtered tibble
}
\description{
Remove isotopocules that are not consistently detected across scans
}
