% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{orbi_filter_weak}
\alias{orbi_filter_weak}
\title{Filter to remove weak isotopocules}
\usage{
orbi_filter_weak(dataset, min_percent)
}
\arguments{
\item{dataset}{A simplified IsoX data frame to be processed}

\item{min_percent}{A number between 0 and 90. Isotopocule must be observed in at least this percentage of scans (please note: the percentage is defined relative to the most commonly observed isotopocule of the compound)}
}
\value{
Filtered tibble
}
\description{
The function \code{orbi_filter_weak()} removes isotopocules that are not consistently detected in most scans
}
\details{
The input \code{dataset} is expected to have at least these 8 columns: \code{filename}, \code{scan.no}, \code{time.min}, \code{compound}, \code{isotopocule}, \code{ions.incremental}, \code{tic}, \code{it.ms}.
}
\examples{
fpath <- system.file("extdata", "testfile_Flow_Exploration_small.isox", package = "isoorbi")
df <- orbi_read_isox(filepath = fpath) \%>\%
                     orbi_simplify_isox() \%>\%
                     orbi_filter_weak(min_percent = 2)

}
