% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{orbi_filter_weak_isotopocules}
\alias{orbi_filter_weak_isotopocules}
\title{Function replaced by \code{orbi_flag_weak_isotopocules()}}
\usage{
orbi_filter_weak_isotopocules(...)
}
\arguments{
\item{...}{parameters passed on to the new function orbi_flag_weak_isotopocules().}
}
\description{
Function replaced by \code{orbi_flag_weak_isotopocules()}
}
