% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{orbi_flag_outliers}
\alias{orbi_flag_outliers}
\title{Flag outlier scans}
\usage{
orbi_flag_outliers(dataset, agc_window)
}
\arguments{
\item{dataset}{Simplified IsoX dataset to have outliers flagged}

\item{agc_window}{flags scans with a critically low or high number of ions in the Orbitrap analyzer. TIC multiplied by injection time serves as an estimate for the number of ions in the Orbitrap. Provide a vector with 2 numbers \code{c(x,y)} flagging the lowest x percent and highest y percent.}
}
\value{
Filtered tibble
}
\description{
The function \code{orbi_flag_outliers()} flags outliers. Grouping is by columns \code{filename} and \code{compound}.
}
\details{
Function is intended to flag scans that are outliers.

The input \code{dataset} is expected to have at least these 8 columns: \code{filename}, \code{scan.no}, \code{time.min}, \code{compound}, \code{isotopocule}, \code{ions.incremental}, \code{tic}, \code{it.ms}.
}
\examples{
fpath <- system.file("extdata", "testfile_flow.isox", package = "isoorbi")
df <-
  orbi_read_isox(file = fpath) |>
  orbi_simplify_isox() |>
  orbi_flag_outliers(agc_window = c(1,99))

}
