% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{orbi_flag_satellite_peaks}
\alias{orbi_flag_satellite_peaks}
\title{Flag minor satellite peaks}
\usage{
orbi_flag_satellite_peaks(dataset)
}
\arguments{
\item{dataset}{A data frame or tibble produced from IsoX data by \code{orbi_simplify_isox()}}
}
\value{
A data frame with new column \code{is_satellite_peak} that flags satellite peaks.
}
\description{
Flag minor signals (e.g., satellite peaks) that were reported by IsoX (filter them out with \code{orbi_filter_flagged_data()}).
}
\details{
The \code{orbi_filter_satellite_peaks()} function removes minor signals for an isotopocule that have been reported by IsoX.
These are often small \verb{satellite peaks} generated by the Fourier transform.

If there are signal of high intensity or very many signals, this can indicate that the m/z and tolerance setting used for processing .raw files with IsoX were incorrect.
}
\examples{
fpath <- system.file("extdata", "testfile_flow.isox", package = "isoorbi")
df <-
  orbi_read_isox(file = fpath) |>
  orbi_simplify_isox() |>
  orbi_flag_satellite_peaks()

}
