% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocks_functions.R
\name{orbi_get_blocks_info}
\alias{orbi_get_blocks_info}
\title{Summarize blocks info}
\usage{
orbi_get_blocks_info(dataset)
}
\arguments{
\item{dataset}{tibble produced by \code{\link[=orbi_define_blocks_for_dual_inlet]{orbi_define_blocks_for_dual_inlet()}}}
}
\description{
FIXME: fully document
}
