% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{orbi_get_settings}
\alias{orbi_get_settings}
\title{Get all isoorbi package settings}
\usage{
orbi_get_settings(pattern = NULL)
}
\arguments{
\item{pattern}{an optional parameter with a regular expression pattern by which to sub-select the returned settings}
}
\value{
list of all package settings and their values
}
\description{
Get all isoorbi package settings
}
\examples{
orbi_get_settings()
}
