% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{orbi_identify_isotopocules}
\alias{orbi_identify_isotopocules}
\title{Identify isotopocules}
\usage{
orbi_identify_isotopocules(aggregated_data, isotopocules)
}
\arguments{
\item{aggregated_data}{either data aggregated from \code{\link[=orbi_aggregate_raw]{orbi_aggregate_raw()}} or a straight-up tibble data frame of the peaks (e.g. retrieved via \code{orbi_get_data(peaks = everything())}).}

\item{isotopocules}{list of isotopocules to map, can be a data frame/tibble or name of a file to read from (.csv/.tsv/.xlsx are all supported).
Required columns are \code{isotopocule/isotopolog}, \code{mz/mass}, \code{tolerance/tolerance [mmu]/tolerance [mDa]}, and \code{charge/z} (these alternative names for the columns, including uppercase versions, are recognized automatically).
Optional column: \verb{#compound/compound} as well as any other columns that don't match these others.
Character columns in the \code{isotopocules} table (including \code{isotopocule} and \code{compound}) are turned into factors with levels that preserve the order of isotopocules.
That means that to change the order of isotopocules in downstream plotting functions, make sure to list them in the order you'd like them presented in.}
}
\value{
same object as provided in \code{aggregated_data} with added columns \code{compound} (if provided), \code{itc_uidx} (introduced unique isotopocule index), \code{isotopocule}, \code{mzExact} and \code{charge} as well as any other additional information columns provided in \code{isotopocules}. Note that the information about columns that were NOT aggregated in previous steps is purposefully not preserved at this step.
}
\description{
Map the mass spectral peaks to specific isotopocules based on their mass.
}
