% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{orbi_identify_isotopocules}
\alias{orbi_identify_isotopocules}
\title{Identify isotopocules}
\usage{
orbi_identify_isotopocules(aggregated_data, isotopocules)
}
\arguments{
\item{aggregated_data}{datasets aggregated from \code{\link[=orbi_aggregate_raw]{orbi_aggregate_raw()}}}

\item{isotopocules}{list of isotopocules to map, can be a data frame/tibble or name of a file (.csv/.tsv/.xlsx are all supported).}
}
\description{
Map the mass spectral peaks to specific isotopocules based on their mass.
}
