% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{orbi_plot_raw_data}
\alias{orbi_plot_raw_data}
\title{Visualize raw data}
\usage{
orbi_plot_raw_data(
  dataset,
  isotopocules,
  x = c("time.min", "scan.no"),
  y = intensity,
  color = "isotopocule",
  colors = c("#1B9E77", "#D95F02", "#7570B3", "#E7298A", "#66A61E", "#E6AB02", "#A6761D",
    "#666666")
)
}
\arguments{
\item{dataset}{isox dataset}

\item{isotopocules}{which isotopocule(s) to visualize}

\item{x}{x-axis for the plot, either "time.min" (the default) or "scan.no"}

\item{y}{expression for what to plot on the y-axis, e.g. \code{intensity}, \code{tic * it.ms}, \code{ratio}, etc. the default is \code{intensity}}

\item{color}{column for the plot, default is "isotopocule"}

\item{colors}{which colors to use, by default a color-blind friendly color palettes (RColorBrewer, dark2)}

\item{add_blocks}{whether to show blocks (if any are defined)}
}
\description{
Call this function to visualize orbitrap data vs. time or scan number. The most common uses are \code{orbi_plot_raw_data(y = intensity)}, \code{orbi_plot_raw_data(y = ratio)}, and \code{orbi_plot_raw_data(y = tic * it.ms)}.
}
