% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{orbi_plot_satellite_peaks}
\alias{orbi_plot_satellite_peaks}
\title{Visualize satellite peaks}
\usage{
orbi_plot_satellite_peaks(
  dataset,
  isotopocules,
  x = c("scan.no", "time.min"),
  colors = c("#1B9E77", "#D95F02", "#7570B3", "#E7298A", "#66A61E", "#E6AB02", "#A6761D",
    "#666666")
)
}
\arguments{
\item{dataset}{isox dataset with satellite peaks identified (\code{orbi_flag_satellite_peaks()})}

\item{isotopocules}{which isotopocules to visualize}

\item{x}{x-axis for the plot, either "time.min" (the default) or "scan.no"}

\item{colors}{which colors to use, by default a color-blind friendly color palettes (RColorBrewer, dark2)}
}
\description{
Call this function any time after flagging the satellite peaks to see where they are. Use the \code{isotopocules} argument to focus on the specific isotopocules of interest.
}
