% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{orbi_plot_satellite_peaks}
\alias{orbi_plot_satellite_peaks}
\title{Visualize satellite peaks}
\usage{
orbi_plot_satellite_peaks(
  dataset,
  isotopocules = c(),
  x = c("scan.no", "time.min"),
  y_scale = c("log", "pseudo-log", "continuous", "none"),
  y_scale_sci_labels = TRUE,
  colors = c("#1B9E77", "#D95F02", "#7570B3", "#E7298A", "#66A61E", "#E6AB02", "#A6761D",
    "#666666"),
  color_scale = scale_color_manual(values = colors)
)
}
\arguments{
\item{dataset}{isox dataset with satellite peaks identified (\code{orbi_flag_satellite_peaks()})}

\item{isotopocules}{which isotopocules to visualize, if none provided will visualize all (this may take a long time or even crash your R session if there are too many isotopocules in the data set)}

\item{x}{x-axis column for the plot, either "time.min" or "scan.no"}

\item{y_scale}{what type of y scale to use: "log" scale, "pseudo-log" scale (smoothly transitions to linear scale around 0), "continuous" scale or "none" if you want to add a y scale to the plot manually instead}

\item{y_scale_sci_labels}{whether to render numbers with scientific exponential notation}

\item{colors}{which colors to use, by default a color-blind friendly color palettes (RColorBrewer, dark2)}

\item{color_scale}{use this parameter to replace the entire color scale rather than just the \code{colors}}
}
\description{
Call this function any time after flagging the satellite peaks to see where they are. Use the \code{isotopocules} argument to focus on the specific isotopocules of interest.
}
