% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isox_data.R
\name{orbi_read_isox}
\alias{orbi_read_isox}
\title{Read IsoX file}
\usage{
orbi_read_isox(filepath)
}
\arguments{
\item{filepath}{Path to the .isox file}
}
\value{
A data frame containing at minimum the columns named \code{filename}, \code{scan.no}, \code{time.min}, \code{compound}, \code{isotopocule}, \code{ions.incremental}, \code{tic}, \code{it.ms}
}
\description{
Read an IsoX output file (.isox) into a tibble
}
\examples{
fpath <- system.file("extdata", "testfile_DualInlet_small.isox", package="isoorbi")
df <- orbi_read_isox(filepath = fpath)

}
