% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_raw.R
\name{orbi_read_raw}
\alias{orbi_read_raw}
\title{Read RAW files}
\usage{
orbi_read_raw(
  file_paths,
  show_progress = rlang::is_interactive(),
  show_problems = TRUE
)
}
\arguments{
\item{file_paths}{paths to the \code{.raw} file(s), single value or vector of paths. Use \code{\link[=orbi_find_raw]{orbi_find_raw()}} to get all raw files in a folder.}

\item{show_progress}{whether to show a progress bar, by default always enabled when running interactively e.g. inside RStudio (and disabled in a notebook), turn off with \code{show_progress = FALSE}}

\item{show_problems}{whether to show problems encountered along the way (rather than just keeping track of them with \code{\link[=orbi_get_problems]{orbi_get_problems()}}). Set to \code{show_problems = FALSE} to turn off the live printout. Either way, all encountered problems can be retrieved with running \code{\link[=orbi_get_problems]{orbi_get_problems()}} for the returned list}
}
\value{
a tibble data frame where each row holds the file path and nested tibbles of datasets extracted from the raw file (typically \code{file_info}, \code{scans}, \code{peaks}, and \code{raw_data}). This is the safest way to extract the data without needing to make assumptions about compatibility across files. Extract your data of interest from the tibble columns or use \code{\link[=orbi_aggregate_raw]{orbi_aggregate_raw()}} to extract safely across files.
}
\description{
Read raw data files (\code{.raw}) from Orbitrap IRMS runs directly. This function extracts all available information and thus can be relatively slow (~1s / Mb on a typical personal computer). The results can be used directly or, more typically, are aggregated with \code{\link[=orbi_aggregate_raw]{orbi_aggregate_raw()}} to safely extract the relevant information for downstream processing.  This function is designed to be fail save by safely catching errors and reporting back on them (see \code{\link[=orbi_get_problems]{orbi_get_problems()}}).
}
