% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_files.R
\name{orbi_read_raw}
\alias{orbi_read_raw}
\title{read raw files with progress bar and summary stats}
\usage{
orbi_read_raw(
  file_paths,
  show_progress = rlang::is_interactive(),
  show_problems = TRUE
)
}
\description{
read raw files with progress bar and summary stats
}
