% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options-package.R
\name{orbi_set_settings}
\alias{orbi_set_settings}
\title{Set package settings}
\usage{
orbi_set_settings(...)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{orbi_set_settings()} was renamed \code{orbi_options()} as part of \code{isoorbi} switching from 'settings' to 'options' to be consistent with base R naming conventions
}
