% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isox_data.R
\name{orbi_simplify_isox}
\alias{orbi_simplify_isox}
\title{Simplify IsoX output}
\usage{
orbi_simplify_isox(dataset)
}
\arguments{
\item{dataset}{The loaded IsoX data that is to be simplified}
}
\value{
A data frame containing only the 8 columns: 'filename', 'scan.no', 'time.min', 'compound', 'isotopolog', 'ions.incremental', 'tic', 'it.ms'.
}
\description{
Keep only columns that are essential for isotopocule ratio analysis
}
\examples{
fpath <- system.file("extdata", "testfile_Flow_Exploration_small.isox", package="isoorbi")
df <- orbi_read_isox(filepath = fpath) \%>\% orbi_simplify_isox()

}
