% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratio_functions.R
\name{orbi_summarize_results}
\alias{orbi_summarize_results}
\title{Generate the results table}
\usage{
orbi_summarize_results(
  dataset,
  ratio_method = c("mean", "sum", "median", "geometric_mean", "slope", "weighted_sum"),
  .by = c("block", "sample_name", "segment", "data_group", "data_type", "injection")
)
}
\arguments{
\item{dataset}{A tibble from \code{IsoX} output (\code{\link[=orbi_read_isox]{orbi_read_isox()}}) and with a basepeak already defined (using \code{orbi_define_basepeak()}). Optionally, with block definitions (\code{\link[=orbi_define_blocks_for_dual_inlet]{orbi_define_blocks_for_dual_inlet()}}) or even additional block segments (\code{\link[=orbi_segment_blocks]{orbi_segment_blocks()}}).}

\item{.by}{additional grouping columns for the results summary (akin to dplyr's \code{.by} parameter e.g. in \code{\link[dplyr:summarise]{dplyr::summarize()}}). If not set by the user, all columns in the parameter's default values are used, if present in the dataset. Note that the order of these is also used to arrange the summary.}
}
\value{
Returns a results summary table retaining the columns \code{filename}, \code{compound}, \code{isotopocule} and \code{basepeak} as well as the grouping columns from the \code{.by} parameter that are part of the input \code{dataset}. Additionally this function adds the following results columns:  \code{ratio}, \code{ratio_sem}, \code{ratio_relative_sem_permil}, \code{shot_noise_permil}, \code{No.of.Scans}, \code{minutes_to_1e6_ions}
\itemize{
\item \code{ratio}: The isotope ratio between the \code{isotopocule} and the \code{basepeak}, calculated using the \code{ratio_method}
\item \code{ratio_sem}: Standard error of the mean for the ratio
\item \code{number_of_scans}: Number of scans used for the final ratio calculation
\item \code{minutes_to_1e6_ions}: Time in minutes it would take to observe 1 million ions of the \code{isotopocule} used as numerator of the ratio calculation.
\item \code{shot_noise_permil}: Estimate of the shot noise (more correctly thermal noise) of the reported ratio in permil.
\item \code{ratio_relative_sem_permil}: Relative standard error of the reported ratio in permil
}
}
\description{
Contains the logic to generate the results table. It passes the  \code{ratio_method} parameter to the \code{\link[=orbi_calculate_ratio]{orbi_calculate_ratio()}} function for ratio calculations.
}
\examples{
fpath <- system.file("extdata", "testfile_flow.isox", package = "isoorbi")
df <- orbi_read_isox(file = fpath) |>
      orbi_simplify_isox() |> orbi_define_basepeak(basepeak_def = "M0")  |>
      orbi_summarize_results(ratio_method = "sum")

}
